[ska] relocate 'shared/lib/security' helpers to 'osquery/cypress' dir (#225438)
## Summary

This PR removes `kbn/test-suites-serverless` as kbn reference for
osquery/cypress test-helper module in order to fix circular dependency I
faced in https://github.com/elastic/kibana/pull/225377

Since the main consumer of that `shared/lib/security` code is
`osquery/cypress` test module, I believe it is a resonable change in
order to unblock us relocating tests from `x-pack/test` dir.

(cherry picked from commit 10cc7b3c62cdd4346792fefa2d403d3ddf759e6a)
