+++ x-pack/platform/plugins/shared/osquery/cypress/lib/default_http_headers.ts
