+++ x-pack/platform/plugins/shared/osquery/cypress/lib/kibana_roles/index.ts
