+++ x-pack/platform/plugins/shared/osquery/cypress/support/setup_data_loader_tasks.ts
@@ -6,11 +6,8 @@
  */
 
 import { createRuntimeServices } from '@kbn/cypress-test-helper/src/services/stack_services';
-import { SecurityRoleAndUserLoader } from '@kbn/test-suites-serverless/shared/lib';
-import type {
-  LoadedRoleAndUser,
-  YamlRoleDefinitions,
-} from '@kbn/test-suites-serverless/shared/lib';
+import { SecurityRoleAndUserLoader } from '../lib';
+import type { LoadedRoleAndUser, YamlRoleDefinitions } from '../lib';
 import type { LoadUserAndRoleCyTaskOptions } from './e2e';
 
 interface AdditionalDefinitions {
