+++ x-pack/platform/plugins/shared/osquery/cypress/tsconfig.json
@@ -5,11 +5,9 @@
     "./cypress.config.ts",
     "./serverless_cypress.config.ts",
     "./serverless_cypress_qa.config.ts",
-    "../../../../../test_serverless/shared/lib",
   ],
   "exclude": [
     "target/**/*",
-    "../../../../../test_serverless/shared/lib/security/default_http_headers.ts"
   ],
   "compilerOptions": {
     "outDir": "target/types",
@@ -20,10 +18,6 @@
     "resolveJsonModule": true,
   },
   "kbn_references": [
-    "@kbn/test-suites-serverless",
-    {
-      "path": "../../../../../test/security_solution_cypress/cypress/tsconfig.json"
-    },
     "@kbn/cypress-config",
     // cypress projects that are nested inside of other ts project use code
     // from the parent ts project in ways that can't be automatically deteceted
@@ -35,5 +29,8 @@
     "@kbn/fleet-plugin",
     "@kbn/cases-plugin",
     "@kbn/cypress-test-helper",
+    "@kbn/security-plugin",
+    "@kbn/test",
+    "@kbn/tooling-log",
   ]
 }
