+++ x-pack/test_serverless/functional/test_suites/security/ftr/discover/context_awareness/cell_renderer.ts
@@ -6,7 +6,6 @@
  */
 
 import expect from '@kbn/expect';
-import { ServerlessRoleName } from '../../../../../../shared/lib';
 import { FtrProviderContext } from '../../../../../ftr_provider_context';
 import { getDiscoverESQLState } from './utils';
 import { SECURITY_SOLUTION_DATA_VIEW, SECURITY_SOLUTION_INDEX_PATTERN } from '../../../constants';
@@ -18,7 +17,7 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
   describe('cell renderer', () => {
     before(async () => {
-      await PageObjects.svlCommonPage.loginWithRole(ServerlessRoleName.PLATFORM_ENGINEER);
+      await PageObjects.svlCommonPage.loginWithRole('platform_engineer');
       await PageObjects.common.navigateToApp('security', {
         path: 'alerts',
       });
