+++ x-pack/test_serverless/shared/lib/index.ts
@@ -5,7 +5,6 @@
  * 2.0.
  */
 
-export * from './security';
 export * from './object_remover';
 export * from './space_path_prefix';
 export * from './cases';
